requirements.txt
# LLM + Embeddings
google-genai>=1.0.0

# Vector DB (local)
chromadb>=0.5.0

# PDF Parsing
pdfplumber>=0.11.0
PyMuPDF>=1.24.0

# Data handling
pydantic>=2.7.0
python-dotenv>=1.0.0
pandas>=2.2.0

# Text processing
tiktoken>=0.7.0

# CLI + Utilities
rich>=13.7.0
typer>=0.12.0
tqdm>=4.66.0
